replace INTEGER with INTEGER*4 to make it easy to write a script to convert them to INTEGER*8 for 64 bit machines

+++ drivers/gidriv.f
@@ -78,7 +78,7 @@ C
 C Note: for 64-bit operating systems, change the following 
 C declaration to INTEGER*8:
 C
-      INTEGER PIXMAP, WORK
+      INTEGER*4 PIXMAP, WORK
 C
       SAVE UNIT, IC, CTABLE, NPICT, MAXIDX, BX, BY, PIXMAP, FILENM
       SAVE CDEFLT, STATE
